Refactor StorageFixture and MotoS3StorageFixtureFactory classes to improve library management and S3 bucket naming conventions; ensure unique bucket names across tests and enhance cleanup logic for shared buckets.

+++ python/arcticdb/storage_fixtures/s3.py
@@ -257,7 +257,10 @@ class BaseS3StorageFixtureFactory(StorageFixtureFactory):
 
     def cleanup_bucket(self, b: S3Bucket):
         # When dealing with a potentially shared bucket, we only clear our the libs we know about:
-        b.slow_cleanup(failure_consequence="We will be charged unless we manually delete it. ")
+        if not self.use_mock_storage_for_testing:
+            # We are not writing to buckets in this case
+            # and if we try to delete the bucket, it will fail
+            b.slow_cleanup(failure_consequence="The following delete bucket call will also fail. ")
 
 
 def real_s3_from_environment_variables(
@@ -631,10 +634,18 @@ class MotoS3StorageFixtureFactory(BaseS3StorageFixtureFactory):
         self.use_raw_prefix = use_raw_prefix
         self.use_mock_storage_for_testing = use_mock_storage_for_testing
         self.use_internal_client_wrapper_for_testing = use_internal_client_wrapper_for_testing
-        # This is needed because we might have multiple factories in the same test
+        # This is needed because we might have multiple factory instances in the same test run
         # and we need to make sure the bucket names are unique
         # set the unique_id to the current UNIX timestamp to avoid conflicts
         self.unique_id = str(int(time.time()))
+
+    def bucket_name(self, bucket_type="s3"):
+        # We need to increment the bucket_id for each new bucket
+        self._bucket_id += 1
+        # We need the unique_id because we have tests that are creating the factory directly
+        # and not using the fixtures
+        # so this guarantees a unique bucket name
+        return f"test_{bucket_type}_bucket_{self.unique_id}_{self._bucket_id}"
 
     def _start_server(self):
         port = self.port = get_ephemeral_port(2)
@@ -724,9 +735,8 @@ class MotoS3StorageFixtureFactory(BaseS3StorageFixtureFactory):
         self._enforcing_permissions = enforcing
 
     def create_fixture(self) -> S3Bucket:
-        bucket = f"test_bucket_{self.unique_id}_{self._bucket_id}"
+        bucket = self.bucket_name("s3")
         self._s3_admin.create_bucket(Bucket=bucket)
-        self._bucket_id += 1
         if self.bucket_versioning:
             self._s3_admin.put_bucket_versioning(Bucket=bucket, VersioningConfiguration={"Status": "Enabled"})
 
@@ -737,7 +747,10 @@ class MotoS3StorageFixtureFactory(BaseS3StorageFixtureFactory):
     def cleanup_bucket(self, b: S3Bucket):
         self._live_buckets.remove(b)
         if len(self._live_buckets):
-            b.slow_cleanup(failure_consequence="The following delete bucket call will also fail. ")
+            if not self.use_mock_storage_for_testing:
+                # We are not writing to buckets in this case
+                # and if we try to delete the bucket, it will fail
+                b.slow_cleanup(failure_consequence="The following delete bucket call will also fail. ")
             try:
                 self._s3_admin.delete_bucket(Bucket=b.bucket)
             except botocore.exceptions.ClientError as e:
@@ -758,9 +771,8 @@ _PermissionCapableFactory = MotoS3StorageFixtureFactory
 
 class MotoNfsBackedS3StorageFixtureFactory(MotoS3StorageFixtureFactory):
     def create_fixture(self) -> NfsS3Bucket:
-        bucket = f"test_bucket_{self.unique_id}_{self._bucket_id}"
+        bucket = self.bucket_name("nfs")
         self._s3_admin.create_bucket(Bucket=bucket)
-        self._bucket_id += 1
         out = NfsS3Bucket(self, bucket)
         self._live_buckets.append(out)
         return out
@@ -800,9 +812,8 @@ class MotoGcpS3StorageFixtureFactory(MotoS3StorageFixtureFactory):
         wait_for_server_to_come_up(self.endpoint, "moto", self._p)
 
     def create_fixture(self) -> GcpS3Bucket:
-        bucket = f"test_bucket_{self.unique_id}_{self._bucket_id}"
+        bucket = self.bucket_name("gcp")
         self._s3_admin.create_bucket(Bucket=bucket)
-        self._bucket_id += 1
         out = GcpS3Bucket(self, bucket)
         self._live_buckets.append(out)
         return out
